Added an Attack/Idle/Walking animations; Added bool to TilePlacement; Added a shield;
Added an Attack/Idle/Walking animations with proper triggers. Still needs a bit of work though. Added bool to TilePlacement to ensure only 1 tower can be placed at a time. Added a shield that is used with 1 handed weapons. Currently has no function but could be use as armor or + X amount of health to the player.