Added some text enemies; Added a next wave countdown timer; Tweaked the enemy spawner;
Added some test enemies to help with testing the wave spawns. Added a next wave countdown timer that is used to count down to the start of the next level. Tweaked the enemy spawner to only spawn X amount per wave. This along with the next wave countdown timer needs a bit more work to function properly.