Added Health Bars; Worked on RepairHammer/ExtraSlot; Created a WeaponController to handle the hitbox collisions;
Added Health Bars to all towers and player that are based off the HealthControllers. Worked on RepairHammer/ExtraSlot so the items occupies the slots. The RepairHammer can also be used to hit towers to repair them. Created a WeaponController to handle the hitbox collisions for all weapons, this might need more work.