Updated the Guide/Pause/Settings Menu GUI; Added a Health Potion; Created a DropOnDeath Script; Created a UsePotion Script; Created a OnGateDestroyed script; Created a UpdatePotionCount script; Had to update the UpdateUIToMatchGold script;
Updated the Guide/Pause/Settings Menu GUI. Guide still needs a bit more information displayed. Added a Health Potion that heals the player for 10HP currently. Created a DropOnDeath Script so enemies can drop potions or other items like a weapon prefab if we want them to. Created a UsePotion Script that lets the user use a potion and heal themselves with the F key. Created a OnGateDestroyed script that sets the isGameOver bool to true inside of the GameManagerLogic script triggering the end of the game. Created a UpdatePotionCount script used to update the textbox that displays the current potion count. Had to update the UpdateUIToMatchGold script. Had to make it Update() instead of FixedUpdate() because it wasn't updating the mene UI properly. I'm guessing its the difference in the rate at which they both update, not 100% on this though.