Changed CameraHolder prefab tag; Hid a few weapons on level3; Edited HealthController; Made a function for global Time.timeScale; Added OnEnabledPause Script; Removed timeScale from LevelUILogic; Edit UI_NextLevel script; Added a few function to GameManagerLogic; Added bool to OnSceneLoaded(); Added gameOver/gameWon checks;
Changed CameraHolder prefab tag and moved "CameraFollower" tag to the CinemachineCamera. Hid a few weapons on level3 for the player to find. Edited HealthController adding a Destroy(gameObject) to the "enemy" if statement. Made a function for global Time.timeScale so every script is using the same one. Might not have needed this but thought it would help an error I had. Added OnEnabledPause Script that pause on enable and unpauses on disable. Removed timeScale from LevelUILogic since it was causing the game to never pause. Edit UI_NextLevel script adding "AddListener" for the button sounds since "RemoveAllListeners" was being used. Added a few function to GameManagerLogic tracking and updated the current scene count that is used to iterate through the listOfRandomSceneOrder list. Added bool to OnSceneLoaded() to be used for checking if the scene has been loaded. Main use is for the "New Game", "Next Level" buttons so they only fire once when clicked. Had and issue where the button click would fire more than once on a click and caused issues with the scene counter; Added gameOver/gameWon checks in UI_NextLevel, UI_NewGame, LevelUILogic making sure things didn't fire at the wrong time.