Updated weapon prefab capsule colliders; Moved around some items on level 2; Made an edit to the enemy spawnDelay;
Updated weapon prefab capsule colliders to make sure they are all on. Moved around some items on level 2 to mix it up a bit adding econ buildings as well. Made an edit to the enemy spawnDelay so after 30 seconds the spawnDelay = .4 to speed things up a bit since it might take a while to spawn everything.